requirements.txt
Flask
pyTelegramBotAPI
openai
python-dotenv
requests
Pillow
